Approved Whip's Warhead Detonator
It is a very simple and short script that's a one-off.
All it does is detonate all of the warheads on the grid when it's run;
it's easy to use with timers or with sensor blocks, and allows for
simultaneous detonation of all warheads without having to pre-arm
your warheads.

- KOMRADE 54B3R